refactor: reorganize application structure and update import paths for consistency
