+++ src/application/email/email-service.interface.ts
