+++ src/application/modules/email.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { EmailServiceToken } from 'src/core/email/email-service.interface';
-import { EmailTransportToken } from 'src/core/email/email-transport.interface';
+import { EmailServiceToken } from 'src/application/email/email-service.interface';
+import { EmailTransportToken } from 'src/application/email/email-transport.interface';
 import { EmailService } from 'src/infrastructure/email/email.service';
 import { NodemailerService } from 'src/infrastructure/email/nodemailer.service';
