+++ src/core/subscription/subscription-service.interface.ts
@@ -1,4 +1,10 @@
-import { SubscriptionPayload } from './subscription-repoository.interface';
+import { Frequency } from './subscription.entity';
+
+export type SubscriptionPayload = {
+  email: string;
+  city: string;
+  frequency: Frequency;
+};
 
 export interface SubscriptionProvider {
   subscribe(payload: SubscriptionPayload): Promise<void>;
